pyproject.toml
[project]
name = "alphagenome-pytorch"
version = "0.2.4"
description = "AlphaGenome"
authors = [
    { name = "Phil Wang", email = "lucidrains@gmail.com" }
]
readme = "README.md"
requires-python = ">= 3.11"
license = { file = "LICENSE" }
keywords = [
    'artificial intelligence',
    'deep learning',
    'transformers',
    'attention mechanism',
    'genomics',
    'splicing',
]

classifiers=[
    'Development Status :: 4 - Beta',
    'Intended Audience :: Developers',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3.11',
]

dependencies = [
    "accelerate",
    "einx>=0.3.0",
    "einops>=0.8.0",
    "PoPE-pytorch>=0.0.9",
    "torch-einops-utils",
    "torch>=2.4",
]

[project.urls]
Homepage = "https://pypi.org/project/alphagenome-pytorch/"
Repository = "https://github.com/lucidrains/alphagenome"

[project.optional-dependencies]
test = [
    "pytest",
    "scipy",
]

example = [
    "pyyaml"
]

convert = [
    "jax[cpu]",
    "dm-haiku",
    "orbax-checkpoint",
    "jmp",
    "einshape",
    "huggingface_hub",
    "alphagenome-research @ git+https://github.com/google-deepmind/alphagenome_research.git",
]

flashattn = [
    "flash-attn",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["alphagenome_pytorch"]
